refactored training code and removed metric functions
