+++ mlops_project/delete/train_model_lightning_old.py
