+++ utils/utils_functions.py
@@ -1,5 +1,6 @@
-import numpy as np
 import os
+from transformers import BertTokenizer
+import torch
 
 
 def read_dataset(tweets_file, labels_file=None):
@@ -48,6 +49,7 @@ def get_datasets():
     return (train_tweets, train_labels), (val_tweets, val_labels), (test_tweets, test_labels)
 
 
+# Preprocessing
 def preprocessing(input_text, tokenizer):
     """
     Returns <class transformers.tokenization_utils_base.BatchEncoding>
@@ -68,43 +70,16 @@ def preprocessing(input_text, tokenizer):
     )
 
 
-# Metrics
-def _b_tp(preds, labels):
-    """Returns True Positives (TP): count of correct predictions of actual class 1"""
-    return sum([preds == labels and preds == 1 for preds, labels in zip(preds, labels)])
+tokenizer = BertTokenizer.from_pretrained("bert-base-uncased")
 
 
-def _b_fp(preds, labels):
-    """Returns False Positives (FP): count of wrong predictions of actual class 1"""
-    return sum([preds != labels and preds == 1 for preds, labels in zip(preds, labels)])
-
-
-def _b_tn(preds, labels):
-    """Returns True Negatives (TN): count of correct predictions of actual class 0"""
-    return sum([preds == labels and preds == 0 for preds, labels in zip(preds, labels)])
-
-
-def _b_fn(preds, labels):
-    """Returns False Negatives (FN): count of wrong predictions of actual class 0"""
-    return sum([preds != labels and preds == 0 for preds, labels in zip(preds, labels)])
-
-
-def b_metrics(preds, labels):
-    """
-    Returns the following metrics:
-      - accuracy    = (TP + TN) / N
-      - precision   = TP / (TP + FP)
-      - recall      = TP / (TP + FN)
-      - specificity = TN / (TN + FP)
-    """
-    preds = np.argmax(preds, axis=1).flatten()
-    labels = labels.flatten()
-    tp = _b_tp(preds, labels)
-    tn = _b_tn(preds, labels)
-    fp = _b_fp(preds, labels)
-    fn = _b_fn(preds, labels)
-    b_accuracy = (tp + tn) / len(labels)
-    b_precision = tp / (tp + fp) if (tp + fp) > 0 else "nan"
-    b_recall = tp / (tp + fn) if (tp + fn) > 0 else "nan"
-    b_specificity = tn / (tn + fp) if (tn + fp) > 0 else "nan"
-    return b_accuracy, b_precision, b_recall, b_specificity
+def tokenize_tweets(tweets):
+    token_ids = []
+    attention_masks = []
+    for sample in tweets:
+        encoding_dict = preprocessing(sample, tokenizer)
+        token_ids.append(encoding_dict["input_ids"])
+        attention_masks.append(encoding_dict["attention_mask"])
+    token_ids = torch.cat(token_ids, dim=0)
+    attention_masks = torch.cat(attention_masks, dim=0)
+    return token_ids, attention_masks
